Remove circular references
