+++ adhocracy4/categories/models.py
@@ -1,6 +1,7 @@
 from django.conf import settings
 from django.db import models
 
+from adhocracy4.categories.form_fields import IconChoiceField
 from adhocracy4.modules import models as module_models
 
 
@@ -22,8 +23,7 @@ class IconField(models.CharField):
         super().contribute_to_class(cls, name, **kwargs)
 
     def formfield(self, **kwargs):
-        from . import forms
-        form_class = kwargs.get('choices_form_class', forms.IconChoiceField)
+        form_class = kwargs.get('choices_form_class', IconChoiceField)
         kwargs['choices_form_class'] = form_class
         return super().formfield(**kwargs)
